requirements.txt
sqlalchemy
pyodbc
pandas
azure-keyvault-secrets
azure-identity
msal
adal